Add: triggers list to the Agent details
