+++ front/components/assistant/details/AssistantDetailsButtonBar.tsx
