+++ front/components/assistant/details/tabs/AgentInfoTab/AssistantKnowledgeSection.tsx
