+++ front/components/assistant/details/tabs/AgentInfoTab/AssistantToolsSection.tsx
