+++ front/pages/w/[wId]/builder/assistants/index.tsx
@@ -19,10 +19,10 @@ import type { InferGetServerSidePropsType } from "next";
 import { useEffect, useMemo, useRef, useState } from "react";
 
 import { AgentEditBar } from "@app/components/assistant/AgentEditBar";
-import { AssistantDetails } from "@app/components/assistant/AssistantDetails";
 import { ConversationsNavigationProvider } from "@app/components/assistant/conversation/ConversationsNavigationProvider";
 import { AssistantSidebarMenu } from "@app/components/assistant/conversation/SidebarMenu";
 import { CreateAgentButton } from "@app/components/assistant/CreateAgentButton";
+import { AssistantDetails } from "@app/components/assistant/details/AssistantDetails";
 import { AssistantsTable } from "@app/components/assistant/manager/AssistantsTable";
 import { TagsFilterMenu } from "@app/components/assistant/TagsFilterMenu";
 import { EmptyCallToAction } from "@app/components/EmptyCallToAction";
